refactor: extract Bullet class to its own module
- Moved Bullet class from sprites/sprites.py to sprites/bullet.py
- Added import for Bullet in sprites/__init__.py
- Verified bullet firing, movement, and collision functionality
