+++ sprites/__init__.py
@@ -1,3 +1,4 @@
 from .sprites import *
 from .starfield import Starfield
 from .meteoroid import Meteoroid
+from .bullet import Bullet
